Move auth e2e tests to test_auth module
